Implement combine generics on SCE (#26).
Combining multiple SCE objects is almost similar to combining Summarized/RangeSummarized Experiments. In combine_rows, the reduced dims and alternative experiment of the first object is present in the combined experiment. For combine_column, we **try** to combine the reduced dimensions and alternative experiments attributes by their names. If they fail, they will be set to `None` in the combined experiment. 

Also added tests and updated docstrings.